Limit allowed duration values to uint16
This simplifies the logic everywhere because no command should ever exceed a uint16 value.
Uint16 has a max value of 65535, which when counting milliseconds amounts to a little more than a minute of constant command.
This should be the upper limit if this ever for some reason gets corrupted or underflows.

+++ API/DeviceControl/ControlLogic.cs
@@ -112,7 +112,7 @@ public static class ControlLogic
             {
                 Id = shockerInfo.Id,
                 RfId = shockerInfo.RfId,
-                Duration = Math.Clamp(shock.Duration, 300, durationMax),
+                Duration = Math.Clamp(shock.Duration, (ushort)300, durationMax),
                 Intensity = Math.Clamp(shock.Intensity, (byte)1, intensityMax),
                 Type = shock.Type,
                 Model = shockerInfo.Model
